@@ -13,26 +13,61 @@
     "npm": "3.10.x"
   },
   "devDependencies": {
-    "babel-cli": "6.x.x",
-    "babel-core": "6.x.x",
-    "babel-eslint": "^7.2.3",
-    "babel-preset-electron": "bengotow/babel-preset-electron#00783dfc438f122997993ae597a41ec315ba121b",
-    "babel-preset-react": "6.x.x",
+    "@types/better-sqlite3": "^5.2.2",
+    "@types/classnames": "^2.2.7",
+    "@types/debug": "^4.1.2",
+    "@types/deep-extend": "^0.4.31",
+    "@types/electron-spellchecker": "^1.1.1",
+    "@types/enzyme": "^3.9.0",
+    "@types/enzyme-adapter-react-16": "^1.0.5",
+    "@types/event-kit": "^2.4.0",
+    "@types/fs-plus": "^3.0.1",
+    "@types/graceful-fs": "^4.1.3",
+    "@types/immutable": "^3.8.7",
+    "@types/is-online": "^8.0.0",
+    "@types/jasmine": "^1.3.1",
+    "@types/keytar": "^4.0.1",
+    "@types/lru-cache": "^5.1.0",
+    "@types/mkdirp": "^0.5.2",
+    "@types/moment": "^2.13.0",
+    "@types/moment-round": "^1.0.3",
+    "@types/moment-timezone": "^0.5.10",
+    "@types/mousetrap": "^1.6.1",
+    "@types/node-emoji": "^1.8.1",
+    "@types/optimist": "^0.0.29",
+    "@types/proxyquire": "^1.3.28",
+    "@types/raven": "^2.5.3",
+    "@types/react": "^16.8.5",
+    "@types/react-color": "^2.14.1",
+    "@types/react-dom": "^16.8.2",
+    "@types/react-test-renderer": "^16.8.1",
+    "@types/react-transition-group": "^1.1.6",
+    "@types/reflux": "^6.4.2",
+    "@types/rimraf": "^2.0.2",
+    "@types/rx-lite": "^4.0.6",
+    "@types/slate": "^0.43.9",
+    "@types/slate-base64-serializer": "^0.2.2",
+    "@types/slate-html-serializer": "^0.6.3",
+    "@types/slate-plain-serializer": "^0.6.1",
+    "@types/slate-react": "^0.20.3",
+    "@types/source-map-support": "^0.4.2",
+    "@types/temp": "^0.8.33",
+    "@types/underscore": "^1.8.9",
+    "@types/underscore.string": "^0.0.33",
+    "@types/uuid": "^3.4.4",
+    "@types/xlsx": "^0.0.36",
+    "@typescript-eslint/eslint-plugin": "^1.4.2",
+    "@typescript-eslint/parser": "^1.4.2",
     "chalk": "1.x.x",
     "devtron": "^1.4.0",
     "electron": "4.0.5",
     "electron-installer-dmg": "0.2.x",
     "electron-packager": "13.0.x",
     "electron-winstaller": "2.x.x",
-    "eslint": "^4.7.2",
-    "eslint-config-prettier": "^2.9.0",
-    "eslint-config-react-app": "^2.0.0",
-    "eslint-plugin-flowtype": "^2.36.0",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^5.1.1",
-    "eslint-plugin-prettier": "^2.5.0",
-    "eslint-plugin-react": "^7.4.0",
-    "eslint_d": "4.2.0",
+    "eslint": "^5.15.0",
+    "eslint-config-prettier": "^4.1.0",
+    "eslint-plugin-react": "^7.12.4",
+    "eslint-plugin-typescript": "^0.14.0",
     "fs-extra": "2.x.x",
     "fs-plus": "2.x.x",
     "gitbook-cli": "^2.3.2",
@@ -59,6 +94,7 @@
     "rimraf": "^2.6.1",
     "targz": "^1.0.1",
     "temp": "0.8.x",
+    "typescript": "^3.3.3333",
     "underscore": "1.8.x"
   },
   "scripts": {
