@@ -1,14 +1,8 @@
 const proxyquire = require('proxyquire');
-const React = require('react');
-const ReactTestUtils = require('react-dom/test-utils');
+import React from 'react';
+import ReactTestUtils from 'react-dom/test-utils';
 
-const {
-  Contact,
-  Message,
-  File,
-  AttachmentStore,
-  MessageBodyProcessor,
-} = require('mailspring-exports');
+import { Contact, Message, File, AttachmentStore, MessageBodyProcessor } from 'mailspring-exports';
 
 class EmailFrameStub extends React.Component {
   render() {
