@@ -1,8 +1,8 @@
-const React = require('react');
-const ReactDOM = require('react-dom');
-const ReactTestUtils = require('react-dom/test-utils');
-const { Message } = require('mailspring-exports');
-const MessageParticipants = require('../lib/message-participants').default;
+import React from 'react';
+import ReactDOM from 'react-dom';
+import ReactTestUtils from 'react-dom/test-utils';
+import { Message } from 'mailspring-exports';
+import MessageParticipants from '../lib/message-participants';
 
 const user_1 = {
   name: 'User One',
@@ -25,7 +25,7 @@ const user_5 = {
   email: 'user5@nylas.com',
 };
 
-const test_message = new Message().fromJSON({
+const test_message = new Message({}).fromJSON({
   id: '111',
   from: [user_1],
   to: [user_2],
