@@ -1,5 +1,6 @@
 import { mount } from 'enzyme';
-import { AccountStore, Account, Actions, React, KeyManager } from 'mailspring-exports';
+import React from 'react';
+import { AccountStore, Account, Actions, KeyManager } from 'mailspring-exports';
 import { ipcRenderer } from 'electron';
 import AccountErrorNotification from '../lib/items/account-error-notif';
 
