@@ -1,4 +1,5 @@
-import { React, PropTypes } from 'mailspring-exports';
+import React from 'react';
+import { PropTypes } from 'mailspring-exports';
 
 export default class MyComposerButton extends React.Component {
   // Note: You should assign a new displayName to avoid naming
